msm: camera: crm: Notify UMD on Error to preform recovery
There can be cases when overflow can occur in KMD due to
many reasons. Some of these errors can be recovered and
needs to be initiated via UMD. Having a notify mechanism
to help UMD to perform recovery during time of such event
will help the device for fast recovery and dump the all
the relevant info regarding the usecase context.

Change-Id: I9d37cb7131c58e40d5bc454eeee8b072a417610d
Signed-off-by: Abhilash Kumar <krabhi@codeaurora.org>